Suppressed 412 exceptions during replication test remote DB create
Fixed the last case of test that didn’t have a unique name.

Given that we uniquely name DBs the 412 will only happen in cases where
we have retried a create because of a network issue. In that case a 412
is a success code because the DB we need for the test exists so we
shouldn’t error.